requirements.txt
# # Core PyTorch and Deep Learning
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0
pytorch-lightning>=2.0.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=9.5.0
albumentations>=1.3.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Geometry and Shape Analysis
shapely>=2.0.0

# Visualization and Plotting
matplotlib>=3.6.0
seaborn>=0.12.0

# Data Handling and I/O
pandas>=2.0.0
h5py>=3.8.0

# Progress Bars and UI
tqdm>=4.65.0
rich>=13.0.0

# Configuration and Utilities
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Development and Testing
pytest>=7.3.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Optional: For enhanced performance
# torchmetrics>=0.11.0
# lightning[extra]>=2.0.0

# Platform-specific optimizations (uncomment as needed)
# For CUDA 11.8
# torch==2.0.1+cu118 torchvision==0.15.2+cu118 -f https://download.pytorch.org/whl/torch_stable.html

# For CUDA 12.1
# torch==2.0.1+cu121 torchvision==0.15.2+cu121 -f https://download.pytorch.org/whl/torch_stable.html

timm>=0.9.0 